+++ src/providers/proxmoxve/cloudconfig.rs
@@ -241,7 +241,7 @@ impl MetadataProvider for ProxmoxVECloudConfig {
                     match dhcp {
                         DhcpSetting::V4 => kargs.push("ip=dhcp".to_string()),
                         DhcpSetting::V6 => kargs.push("ip=dhcp6".to_string()),
-                        DhcpSetting::Both => kargs.push("ip=dhcp,dhcp6".to_string()),
+                        DhcpSetting::Both => kargs.push("ip=any".to_string()),
                     }
                 }
 
@@ -403,10 +403,16 @@ impl ProxmoxVECloudNetworkConfigEntry {
             }
 
             if subnet.subnet_type == "dhcp" || subnet.subnet_type == "dhcp4" {
-                iface.dhcp = Some(DhcpSetting::V4)
+                iface.dhcp = iface
+                    .dhcp
+                    .map(|d| d.merge(DhcpSetting::V4))
+                    .or(Some(DhcpSetting::V4))
             }
             if subnet.subnet_type == "dhcp6" {
-                iface.dhcp = Some(DhcpSetting::V6)
+                iface.dhcp = iface
+                    .dhcp
+                    .map(|d| d.merge(DhcpSetting::V6))
+                    .or(Some(DhcpSetting::V6))
             }
             if subnet.subnet_type == "ipv6_slaac" {
                 warn!("subnet type \"ipv6_slaac\" not supported, ignoring");