+++ docs/release-notes.md
@@ -12,6 +12,7 @@ Major changes:
 - Azure: Add `render-ignition` subcommand to generate Ignition config fragments from IMDS metadata
 - Hetzner: Add support for network configuration
 - Support the generation of NetworkManager connection profiles (keyfile format)
+- Support systemd-network-generator by emitting `/run/afterburn/network-generator.env` from the initrd
 
 Minor changes:
 
@@ -24,6 +25,10 @@ Minor changes:
 - Drop the maplit dependency
 - Don't panic truncating an over-long hostname with multi-byte characters
 - ProxmoxVE: Emit one `nameserver=` karg per address so multiple DNS servers work
+- ProxmoxVE: Prevent rd-network-kargs from failing without a config drive
+- KubeVirt: Use `off` rather than `static` to disable IP autoconf for Dracut and systemd compatibility
+- KubeVirt, ProxmoxVE: Use `ip=any` rather than `ip=dhcp,dhcp6` for systemd compatibility. As of Dracut 112, its network-legacy module treats `ip=any` as IPv4-only, so this change breaks IPv6 in that case. At the time of writing, a Dracut change to make this dual stack is under review.
+- ProxmoxVE: Fix dual stack network configuration
 
 Packaging changes: